finally works with db implementation

+++ de.uni_koeln.spinfo.strings.core/src/de/uni_koeln/spinfo/strings/algo/suffixtrees/node/Node.java
@@ -1,12 +1,11 @@
 package de.uni_koeln.spinfo.strings.algo.suffixtrees.node;
 
-import java.util.HashMap;
 
 public interface Node {
 
 	public static final int A_LEAF = -1;
 	
-	int getId();
+	long getId();
 	
 	int getTextNumber();
